error persists despite moving from impls into scope of macro expansion
